# [Xamarin.Android.Build.Tasks] use typemaps when EmbedAssembliesIntoApk=False