--- a/.external
+++ b/.external
@@ -1,2 +1,2 @@
-xamarin/monodroid:master@6f77f153e386c22efb3b508b270e5cdc3dfced61
+xamarin/monodroid:master@ab054acc3fef6ae979c8dd447fb9723eee83c9de
 mono/mono:2019-08@df5e13f95df7a2d11d86904e74b1bd8950c9d43b