@grendello 2019-09-26
@jonathanpeppers out of curiosity, what was the crash you got that this PR fixes?