@jonathanpeppers 2019-09-26
@grendello oh the crash was when I was trying to get this to work, sorry. It was before I made the right `typemap.*.s` changes.