@jonathanpeppers 2019-10-17
The only two test failures are `zipalign` on macOS and one related to `lint` on Windows. These should be unrelated, so merging.